Company-Solutions/customer-support-bot/requirements.txt
fastapi
uvicorn
openai
python-dotenv
faiss-cpu
sentence-transformers
neo4j
